+++ dbcat/catalog/__init__.py
@@ -1,0 +1,4 @@
+# flake8: noqa
+
+from .catalog import Catalog
+from .models import CatColumn, CatSchema, CatSource, CatTable, ColumnLineage
