+++ dbcat/dbcat.py
@@ -1,13 +1,6 @@
-import yaml
-
 from dbcat.catalog.catalog import Catalog
 from dbcat.catalog.db import DbScanner
 from dbcat.log_mixin import LogMixin
-
-
-def catalog_connection(config: str) -> Catalog:
-    config_yaml = yaml.safe_load(config)
-    return Catalog(**config_yaml["catalog"])
 
 
 def pull(catalog: Catalog) -> None:
